# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Palplus UAA",
        "description": "Palplus Swagger API description",
        "contact": {
            "email": "michaelkinare@gmail.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/",
    "paths": {
...
        "/group": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "Retrieve Groups",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "Create Group",
                "parameters": [
                    {
                        "name": "type_id",
                        "in": "query",
                        "description": "group type id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Group Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Group Description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "avatar",
                        "in": "query",
                        "description": "avatar",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "access_level",
                        "in": "query",
                        "description": "access level",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "country",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "description": "currency",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/group/{id}": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "Retrieve Group",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "group id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Group"
                ],
                "summary": "Delete Group",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "group id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Group"
                ],
                "summary": "Update Group",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "group id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Group Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Group Description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "access_level",
                        "in": "query",
                        "description": "access level",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/group/{id}/force": {
            "delete": {
                "tags": [
                    "Group"
                ],
                "summary": "Force delete Group",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "group id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/group/join/{user_id}/{group_id}": {
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "Join group",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "user id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "description": "Group Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/group/leave/{member_id}/{group_id}": {
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "Leave group",
                "parameters": [
                    {
                        "name": "member_id",
                        "in": "query",
                        "description": "Member_id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "description": "Group Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/group/make-admin/{member_id}/{group_id}": {
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "Leave group",
                "parameters": [
                    {
                        "name": "member_id",
                        "in": "query",
                        "description": "Member_id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "description": "Group Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/group/revoke-admin/{member_id}/{group_id}": {
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "Leave group",
                "parameters": [
                    {
                        "name": "member_id",
                        "in": "query",
                        "description": "Member_id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "description": "Group Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/group/members/{group_id}": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "Group members",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "description": "Group Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "description": "Auth Bearer Token Format as 'Bearer <access_token>'",
            "name": "Authorization",
            "in": "header"
        }
    }
}